backend/pyproject.toml
[project]
name = "polyscoop-backend"
version = "0.1.0"
description = "polyscoop – Polymarket data dashboard backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]",
    "uvicorn[standard]",
    "pydantic-settings",
    "httpx",
    "python-jose[cryptography]",
    "passlib[bcrypt]",
    "py-clob-client",
    "python-dotenv",
    "ruff",
    "pyright",
    "pytest",
    "asyncpg",
    "sqlalchemy[asyncio]",
    "alembic",
    "redis[hiredis]",
    "websockets",
    "psycopg2-binary>=2.9.11",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pyright]
pythonVersion = "3.12"
typeCheckingMode = "standard"
venvPath = "."
venv = ".venv"

[tool.pytest.ini_options]
testpaths = ["tests"]
